<feat/contenPosting>: two-check likecount & comment count

+++ src/main/java/com/api/dto/response/ContentPostingResponse.java
@@ -16,7 +16,7 @@ public class ContentPostingResponse {
     private List<Map<String, String>> categories;
     private ZonedDateTime createdDate;
     private boolean isPublic;
-    private int commentCount;
+    private long commentCount;
     private long likeCount;
 
     public ContentPostingResponse() {
@@ -24,7 +24,7 @@ public class ContentPostingResponse {
 
     public ContentPostingResponse(String contentId, String contentName, String userId, String userName,
             String userAvatar, String content, String imageUrl, List<Map<String, String>> categories,
-            ZonedDateTime createdDate, boolean isPublic, int commentCount, long likeCount) {
+            ZonedDateTime createdDate, boolean isPublic, long commentCount, long likeCount) {
         this.contentId = contentId;
         this.contentName = contentName;
         this.userId = userId;
@@ -127,11 +127,11 @@ public class ContentPostingResponse {
         this.isPublic = isPublic;
     }
 
-    public int getCommentCount() {
+    public long getCommentCount() {
         return commentCount;
     }
 
-    public void setCommentCount(int commentCount) {
+    public void setCommentCount(long commentCount) {
         this.commentCount = commentCount;
     }
